+++ lib/custom_notification_scheduler_method_channel.dart
@@ -1,17 +1,12 @@
-import 'package:flutter/foundation.dart';
+import 'package:custom_notification_scheduler/custom_notification_scheduler_platform_interface.dart';
 import 'package:flutter/services.dart';
 
-import 'custom_notification_scheduler_platform_interface.dart';
-
-/// An implementation of [CustomNotificationSchedulerPlatform] that uses method channels.
 class MethodChannelCustomNotificationScheduler extends CustomNotificationSchedulerPlatform {
-  /// The method channel used to interact with the native platform.
-  @visibleForTesting
-  final methodChannel = const MethodChannel('custom_notification_scheduler');
+  static const MethodChannel _channel = MethodChannel('app.vercel.zihadsikder.custom_notification_scheduler');
 
   @override
   Future<String?> getPlatformVersion() async {
-    final version = await methodChannel.invokeMethod<String>('getPlatformVersion');
+    final version = await _channel.invokeMethod<String>('getPlatformVersion');
     return version;
   }
 }